Refactored vnaScanMultithreaded.c
https://stgit.dcs.gla.ac.uk/team-project-h/2025/jh05/jh05-main/-/issues/50

Reworked many parts of the program for clarity and efficiency. Particularly
 moved scan handling and thread coordination from main() to scan_coordinato
r(), and reworked the global variables for port settings restoration and er
ror handling to allow for multiple VNAs and to avoid data duplication -- sh
ould now also be much easier to understand.

Work is still needed to make the global variables used for multithreading sa
fer -- likely by creating and destroying them within scan_coordinator and po
tentially passing pointers to them as arguments (but that may be memory inef
ficient, it might be best to stick to global variables). At the moment those
 variables would not be reset between scans etc.