create_bounded_buffer returns errors, destroy frees main struct
https://stgit.dcs.gla.ac.uk/team-project-h/2025/jh05/jh05-main/-/merge_requests/1#note_354127

Refactored create_bounded_buffer so that it works on a pointer to the
bounded buffer and returns int errors. Also fixed destroy bounded buffer
so that it properly gets rid of this structure without having to free
the malloced space (technically not necessary to put the Bounded Buffer
on the heap but seems strange not to).