Restructure find_binary_header to pull maximum points at a time
https://stgit.dcs.gla.ac.uk/team-project-h/2025/jh05/jh05-main/-/issues/85

find_binary_header now takes in an extra argument, a pointer to a
malloced memory address in which to store the first point of a pull. It
takes in data in chunks of 20 bytes instead of one, which should
significantly reduce the number of system calls and thus the amount of
time spent waiting on input during the find_binary_header stage.

It cannot take more than 20 bytes because if it did it might skip over
and take data from the first two points, which would be an enormous
issue to deal with and take far more flagging than it is worth. This is
the best conmpromise I think.