Replaced POINTS macro with global variable
https://stgit.dcs.gla.ac.uk/team-project-h/2025/jh05/jh05-main/-/issues/86

Can now set POINTS when calling run_multithreaded_scan (passed in
through CLI arguments in VnaScanMultithreadedMain.c). This has the
side-effect that the size of the array of nanovna_raw_datapoint
measurements within the struct datapoint_nanoVNA_H cannot be known at
compile time, so instead of having a hard-coded array the struct
datapoint_nanoVNA_H now has a pointer to an array of raw datapoints.

This introduces more mallocs and frees, and probably worsens data
locality, but I don't think the performance issue should be massive and
this functionality was a request from the customer.