Seperated out main function and made arguments global variables
https://stgit.dcs.gla.ac.uk/team-project-h/2025/jh05/jh05-main/-/issues/94

Main function now seperate, just exists as an entry function to call
readCommand. Much better style and make everything much more testable.
Arguments now global to avoid passing around copies of them everywhere,
as it should be very clear how they are changed (pretty much just setter
functions).
We will need to be careful to follow good style when it comes to
accessing the global variables, as C does not enforce this (but we want
to kind of pretend we're using an OO language lol)